Leaving section
By the agreement, you have a grace period specified by law. If you agree and terminate before that specific time, you will be required to work on an opensource project for the remaining time since the end of the year.
If you're not committing and reporting on the opensource or social project, mondora will ask you to refund this.  This money will be used for opensource projects.